Add patch to fix gcc-10 FTBFS (closes: #957943).
A number of global variables are declared in headers with no explicit
linkage.  This results in there being definitions of them in multiple
object files and causes linking to fail under gcc-10.

Add `extern` to the header declarations and separate declarations with
no linkage in files where they are assigned.